Rework button management for ClickRelease & HoldRelease
Added HoldRelease as a new button event to allow use
of hold and long hold events without conflict. Also added
LongHoldRelease, but will probably continue to just
use LongHold as it is easier on the user.

Adjusted the the milliseconds for various holds and clicks:
ClickRelease is press and release with 500ms
HoldRelease: Count to two and let go of button
LongHold: Count to 3 and let go of button

Reverted prior mod that added CH8 toggle since CH8 is
used for raw gimbal paddle in smart shots.

Filtered out button events not used on the solo from
being sent to the solo so they don't clog up the solo's
button handling.